Add files via upload
Task elemnts changed:
-Paginating
-Sorting
-Filtering

+++ router/tasks.js
@@ -1,10 +1,9 @@
 //modules
 const express = require('express')
 const auth = require('../middleware/auth')
+const router = new express.Router()
 //models
 const Task = require('../models/tasks')
-
-const router = new express.Router()
 
 //create data//
 router.post('/tasks',auth,async (req,res)=>{
@@ -21,17 +20,33 @@ router.post('/tasks',auth,async (req,res)=>{
 })
 
 //read data//
-
 router.get('/tasks', auth,async (req,res)=>{
+    const match = {}
+    const sort = {}
+    if(req.query.completed){
+        match.completed = req.query.completed === 'true'
+    }
+    if(req.query.sortBy){
+        const parts = req.query.sortBy.split(':')
+        sort[parts[0]] = parts[1] === 'asc' ? 1 : -1 
+    }
+    
     try{
-        await req.user.populate('tasks').execPopulate()
+        await req.user.populate({
+            path: 'tasks',
+            match,
+            options:{
+                limit: parseInt(req.query.limit),
+                skip: parseInt(req.query.skip),
+                sort
+            }
+        }).execPopulate()
         //const tasks = await Task.find({owner: req.user._id})
         res.send(req.user.tasks)
     }catch(e){
         res.status(500).send("Error on tasksReading")
     }
 })
-
 router.get('/tasks/:id', auth,async (req,res)=>{
     const _id = req.params.id
     try{
@@ -44,13 +59,11 @@ router.get('/tasks/:id', auth,async (req,res)=>{
     }catch(e){
         res.status(500).send("Error on reading tasks")
     }
-    
 
 })
 
 //updating data
-
-router.patch('/task/:id', auth, async(req, res)=>{
+router.patch('/tasks/:id', auth, async(req, res)=>{
     const updates = Object.keys(req.body)
     const allowedUpdates = ['name', 'description', 'completed']
     const isValidUpdate = updates.every((update)=> allowedUpdates.includes(update))
@@ -70,12 +83,12 @@ router.patch('/task/:id', auth, async(req, res)=>{
         
     }
 })
-//deleting data
 
-router.delete('/tasks', async(req, res)=>{
+//deleting data
+router.delete('/tasks', auth, async(req, res)=>{
     
     try{
-        const tasks = Task.findOneAndDelete({})
+        const tasks = Task.findOneAndDelete({owner: req.user._id})
         
         res.status(200).send('Deleted all files')
     }catch(e){
@@ -83,12 +96,12 @@ router.delete('/tasks', async(req, res)=>{
     }
 
 })
-router.delete('/tasks/:id', async(req,res)=>{
+router.delete('/tasks/:id',auth, async(req,res)=>{
 
     try{
-        const task = await Task.findByIdAndDelete(req.params.id)
+        const task = await Task.findOneAndDelete({_id: req.params.id, owner: req.user._id})
         if(!task){ return res.status(404).send("Error: the task by id:", req.params.id, "do not exist")}
-        res.status(200).send("the task: ", task, "has been deleted")
+        res.send("the task: ", task, "has been deleted")
     }catch(e){
         res.status(500).send()
     }
